feat: refactoring reducers to operators
Moving group reducer to groupBy operator for simplicity
Moving projections operators to mappers
Implementing "type-check" file to allow compile time checks

+++ src/operators/pop/pop.ts
@@ -2,6 +2,10 @@ import { Collection } from '../../types';
 import { isArray, isCollection } from '../../utils';
 
 const collectionPopImpl = (collection: Collection) => {
+  if (isArray(collection)) {
+    return collection.pop();
+  }
+
   const result = [...collection].pop();
   collection.delete([...collection.keys()].pop());
   return result;
@@ -10,4 +14,4 @@ const collectionPopImpl = (collection: Collection) => {
 export const pop =
   () =>
   <S>(source: S) =>
-    isArray(source) ? source.pop() : isCollection(source) ? collectionPopImpl(source) : source;
+    isCollection(source) ? collectionPopImpl(source) : source;
