+++ src/operators/shift/shift.ts
@@ -2,6 +2,10 @@ import { Collection } from '../../types';
 import { isArray, isCollection } from '../../utils';
 
 const collectionShiftImpl = (collection: Collection) => {
+  if (isArray(collection)) {
+    return collection.shift();
+  }
+
   const result = [...collection].shift();
   collection.delete([...collection.keys()].shift());
   return result;
@@ -10,4 +14,4 @@ const collectionShiftImpl = (collection: Collection) => {
 export const shift =
   () =>
   <S>(source: S) =>
-    isArray(source) ? source.shift() : isCollection(source) ? collectionShiftImpl(source) : source;
+    isCollection(source) ? collectionShiftImpl(source) : source;
